Added main.py and data.json
These are example documents of how to use the library.
Running main.py requires data.json and knn.py in the same folder. It is a simple machine learning program that takes a color hex code and predicts if it is black, white, red, blue, or green. In other words, which of the primary colors (including white and black).